base.js is moved to fixture folder
