+++ tests/navigation.spec.js
@@ -1,5 +1,5 @@
 import { expect } from '@playwright/test';
-import { test, loginBusinessUser } from "./base.js";
+import { test, loginBusinessUser } from "../fixtures/base.js";
 import { URL_END_POINTS } from '../testData.js';
 
 const BASE_URL = process.env.URL;
@@ -11,6 +11,6 @@ test('Navigate to login page', async ({ page }) => {
 });
 
 test('Check login fixture', async ({ page, loginBusinessUser}) => { 
-  
+
   await expect(page).toHaveURL(BASE_URL + URL_END_POINTS.signEndPoint);
 });
